+++ inc/core/lxUnicode.h
@@ -15,7 +15,7 @@ The passed string must be nul-terminated.
 \param [out] outBytes If not null the number of bytes in the string is written here
 \return The number of characters in the string without NUL.
 */
-LUX_API size_t StringLengthUTF8(const char* str, size_t* outBytes = nullptr);
+LUX_API int StringLengthUTF8(const char* str, int* outBytes = nullptr);
 
 //! Moves the passed utf8 cursor onto the character before the current one
 LUX_API void RetractCursorUTF8(const char*& ptr);
@@ -40,7 +40,7 @@ The passed string must be nul-terminated.
 \param [out] outBytes If not null the number of bytes in the string is written here
 \return The number of characters in the string without NUL.
 */
-LUX_API size_t StringLengthUTF16(const char* str, size_t* outBytes = nullptr);
+LUX_API int StringLengthUTF16(const char* str, int* outBytes = nullptr);
 
 //! Moves the passed utf16 cursor onto the character after the current one
 /**
