+++ inc/scene/particle/ParticleSystemComponent.h
@@ -20,9 +20,9 @@ public:
 
 	LUX_API void AddTemplate(ParticleSystemTemplate* templ);
 	LUX_API void RemoveTemplate(ParticleSystemTemplate* templ);
-	LUX_API size_t GetTemplateCount() const;
-	LUX_API ParticleSystemTemplate* GetTemplate(size_t i);
-	LUX_API const ParticleSystemTemplate* GetTemplate(size_t i) const;
+	LUX_API int GetTemplateCount() const;
+	LUX_API ParticleSystemTemplate* GetTemplate(int i);
+	LUX_API const ParticleSystemTemplate* GetTemplate(int i) const;
 	LUX_API void VisitRenderables(RenderableVisitor* visitor, scene::ERenderableTags tags);
 	LUX_API void Animate(float time);
 	LUX_API void Render(Node* node, video::Renderer* renderer, const SceneData& sceneData);
